requirements.txt
Flask>=3.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
requests>=2.28.0
python-dotenv>=1.0.0
joblib>=1.0.1
Werkzeug>=2.0.3
beautifulsoup4>=4.9.3
geopy>=2.2.0
Pillow>=8.4.0 